moves config / settings
