+++ opsml/app/core/event_handlers.py
@@ -9,13 +9,13 @@ import rollbar
 from fastapi import FastAPI, Response
 from sqlalchemy.engine.base import Engine
 
-from opsml.app.core.config import config
 from opsml.helpers.logging import ArtifactLogger
 from opsml.registry.model.registrar import ModelRegistrar
 from opsml.registry.sql.db_initializer import DBInitializer
 from opsml.registry.sql.registry import CardRegistries
 from opsml.registry.sql.table_names import RegistryTableNames
-from opsml.registry.utils.settings import settings
+from opsml.registry.storage.settings import settings
+from opsml.settings.config import config
 
 logger = ArtifactLogger.get_logger()
