+++ opsml/app/routes/healthcheck.py
@@ -3,8 +3,8 @@
 # LICENSE file in the root directory of this source tree.
 from fastapi import APIRouter, HTTPException
 
-from opsml.app.core.config import config
 from opsml.app.routes.pydantic_models import DebugResponse, HealthCheckResult
+from opsml.settings.config import config
 
 router = APIRouter()
