+++ opsml/registry/sql/base/__init__.py
@@ -2,7 +2,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from opsml.registry.utils.settings import settings
+from opsml.registry.storage.settings import settings
 
 
 # no typing in order to prevent imports before they're needed
